Refactored initial code
